feat: LoadingBox, UserAvatar, Pagination 공통 컴포넌트 분리 (#33)
- LoadingBox: Spinner 래퍼 컴포넌트 추가 (size/label/className props)
- UserAvatar: 프로필 이미지 + 이니셜 fallback 컴포넌트 추가
- Pagination: QnaListPage 인라인 페이지네이션 추출
- common/index.ts: 3개 신규 컴포넌트 barrel export 추가
- AnswerCard: 인라인 아바타 블록 → UserAvatar 컴포넌트 사용
- QnaListPage: Spinner 인라인 → LoadingBox, 인라인 함수 → Pagination 사용
