+++ src/components/common/LoadingBox/index.ts
@@ -1,0 +1,1 @@
+export * from './LoadingBox'
