+++ src/components/common/Pagination/index.ts
@@ -1,0 +1,1 @@
+export * from './Pagination'
