+++ src/components/common/UserAvatar/index.ts
@@ -1,0 +1,1 @@
+export * from './UserAvatar'
