+++ src/components/common/index.ts
@@ -5,7 +5,9 @@ export * from './Card'
 export * from './Checkbox'
 export * from './Dropdown'
 export * from './Input'
+export * from './LoadingBox'
 export * from './Modal'
+export * from './Pagination'
 export * from './PasswordInput'
 export * from './SearchInput'
 export * from './SocialLoginButton'
@@ -13,3 +15,4 @@ export * from './Spinner'
 export * from './SuccessCard'
 export * from './Tabs'
 export * from './Toast'
+export * from './UserAvatar'
