+++ src/pages/qna/QnaListPage.tsx
@@ -11,7 +11,8 @@ import {
   SearchInput,
   Modal,
   Button,
-  Spinner,
+  LoadingBox,
+  Pagination,
   QuestionCard,
   CategoryFilter,
 } from '@/components'
@@ -31,105 +32,6 @@ const SEARCH_DEBOUNCE_MS = 300
 const SORT_LABEL: Record<SortOption, string> = {
   latest: '최신순',
   oldest: '오래된순',
-}
-
-// ── 페이지네이션 ──────────────────────────────────────────────────────
-
-function ChevronLeft() {
-  return (
-    <svg
-      width="16"
-      height="16"
-      viewBox="0 0 16 16"
-      fill="none"
-      aria-hidden="true"
-    >
-      <path
-        d="M10 12L6 8L10 4"
-        stroke="currentColor"
-        strokeWidth="1.5"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-    </svg>
-  )
-}
-
-function ChevronRight() {
-  return (
-    <svg
-      width="16"
-      height="16"
-      viewBox="0 0 16 16"
-      fill="none"
-      aria-hidden="true"
-    >
-      <path
-        d="M6 4L10 8L6 12"
-        stroke="currentColor"
-        strokeWidth="1.5"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-    </svg>
-  )
-}
-
-function Pagination({
-  current,
-  total,
-  onChange,
-}: {
-  current: number
-  total: number
-  onChange: (page: number) => void
-}) {
-  const maxVisible = 5
-  const half = Math.floor(maxVisible / 2)
-  const start = Math.max(1, Math.min(current - half, total - maxVisible + 1))
-  const end = Math.min(total, start + maxVisible - 1)
-  const pages = Array.from({ length: end - start + 1 }, (_, i) => start + i)
-
-  return (
-    <nav
-      aria-label="페이지 이동"
-      className="mt-8 flex items-center justify-center gap-1"
-    >
-      <button
-        onClick={() => onChange(current - 1)}
-        disabled={current === 1}
-        aria-label="이전 페이지"
-        className="text-text-muted hover:text-text-heading flex h-9 w-9 items-center justify-center rounded-full transition-colors disabled:cursor-not-allowed disabled:text-gray-300"
-      >
-        <ChevronLeft />
-      </button>
-
-      {pages.map((p) => (
-        <button
-          key={p}
-          onClick={() => onChange(p)}
-          aria-current={p === current ? 'page' : undefined}
-          className={[
-            'h-9 w-9 rounded-full text-sm font-medium transition-colors',
-            p === current
-              ? 'bg-primary text-text-inverse'
-              : 'text-text-body hover:bg-bg-muted',
-          ].join(' ')}
-        >
-          {p}
-        </button>
-      ))}
-
-      <button
-        onClick={() => onChange(current + 1)}
-        disabled={current === total}
-        aria-label="다음 페이지"
-        className="text-text-muted hover:text-text-heading flex h-9 w-9 items-center justify-center rounded-full transition-colors disabled:cursor-not-allowed disabled:text-gray-300"
-      >
-        <ChevronRight />
-      </button>
-    </nav>
-  )
 }
 
 // ── QnaListPage ───────────────────────────────────────────────────
@@ -337,9 +239,11 @@ export function QnaListPage() {
 
           {/* 질문 목록 */}
           {isLoading && (
-            <div className="flex items-center justify-center py-20">
-              <Spinner size="lg" label="질문 목록을 불러오는 중..." />
-            </div>
+            <LoadingBox
+              size="lg"
+              label="질문 목록을 불러오는 중..."
+              className="py-20"
+            />
           )}
 
           {isError && (
@@ -391,9 +295,7 @@ export function QnaListPage() {
       >
         <Suspense
           fallback={
-            <div className="flex justify-center py-8">
-              <Spinner label="카테고리 불러오는 중..." />
-            </div>
+            <LoadingBox label="카테고리 불러오는 중..." className="py-8" />
           }
         >
           <CategoryFilter
